Added new .ccp and .h files to follow Craigs naming conventions
