+++ Lab2_Grading/CStudent.h
@@ -1,0 +1,17 @@
+#pragma once
+#ifndef CSTUDENT_H
+#define CSTUDENT_H
+
+class CStudent
+{
+public:
+    std::string     student_Num;
+    float           lab_Grade;
+    float           quiz_Grade;
+    float           midterm_Grade;
+    float           finalExam_Grade;
+};
+
+#endif // !1
+
+
